# docs/openapi.yaml
...
openapi: 3.0.4
info:
  title: dployr Base API
  version: 1.0.0
  description: |
    Your app, your server, your rules!

    ## Authentication
    Authenticate via OAuth2.0 (Google, GitHub, Microsoft) or email OTP.
    Sessions are cookie-based.

    ## Authorization
    Role-based access control:
    - **Owner**: Full control, manages admins and billing
    - **Admin**: User management, infrastructure
    - **Developer**: Deploy apps, manage integrations
    - **Viewer**: Read-only access

    ## Rate Limiting
    - Global: 100 requests/minute per user or IP on `/v1/*`
    - Strict: 10 requests/minute on sensitive endpoints
    - Returns `429` with `Retry-After` and `X-RateLimit-*` headers on breach
servers:
  - url: https://base.dployr.dev
    description: Development
  - url: https://base.dployr.io
    description: Production
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
              example: Resource not found
            code:
              type: string
              description: |
                Machine-readable error code. Known values:
                - `request.missing_params`
                - `request.bad_request`
                - `request.invalid_otp`
                - `request.unprocessable_entity`
                - `request.too_many_requests`
                - `request.internal_server_error`
                - `validation.missing_fields`
                - `validation.invalid_format`
                - `auth.bad_session`
                - `auth.bad_token`
                - `auth.invalid_oauth_provider`
                - `auth.bad_oauth_state`
                - `permission.forbidden`
                - `permission.owner_role_required`
                - `permission.admin_role_required`
                - `permission.developer_role_required`
                - `permission.viewer_role_required`
                - `resource.missing_resource`
                - `resource.conflict`
                - `runtime.bad_webhook_signature`
                - `runtime.internal_server_error`
                - `runtime.node_not_connected`
                - `bootstrap.bootstrap_setup_failure`
                - `bootstrap.bootstrap_run_failure`
              example: resource.missing_resource
            helpLink:
              type: string
              format: uri
              example: https://status.dployr.io
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          nullable: true
          type: object
...
paths:
...
  /v1/node/token:
    post:
      operationId: nodeRefreshToken
      tags: [Nodes]
      summary: Exchange node token for a fresh short-lived token
      description: |
        Called by the daemon to refresh its access token.
        Accepts an existing bootstrap or node token (may be expired for bootstrap type).
      security:
        - bearerAuth: []
      responses:
        '200':
          description: New token
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          token:
                            type: string
        '401':
          description: Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  message: Invalid token
                  code: auth.bad_token
        '404':
          description: Instance not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/proxy/node/routes:
    post:
      operationId: addProxyNodeRoute
      tags: [Proxy]
      summary: Add a proxy route on a node (fire-and-return)
      description: |
        Dispatches a proxy route add task to the target node and returns immediately.
        Requires developer role on the cluster.
      security:
        - sessionCookie: []
      parameters:
        - in: query
          name: clusterId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [instanceName, serviceName, upstream]
              properties:
                instanceName:
                  type: string
                  description: Target instance tag/name
                serviceName:
                  type: string
                upstream:
                  type: string
                  description: Upstream address (host:port)
                domain:
                  type: string
                  nullable: true
                  description: Optional custom domain
                template:
                  type: string
                  nullable: true
                  description: Optional proxy template name
      responses:
        '202':
          description: Route add task dispatched
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          taskId:
                            type: string
                          message:
                            type: string
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions (developer role required)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Node not connected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/proxy/node/routes/{service}:
    delete:
      operationId: deleteProxyNodeRoute
      tags: [Proxy]
      summary: Delete a proxy route on a node (fire-and-return)
      description: |
        Dispatches a proxy route removal task to the target node and returns immediately.
        Requires developer role on the cluster.
      security:
        - sessionCookie: []
      parameters:
        - in: path
          name: service
          required: true
          schema:
            type: string
          description: Service name whose route should be removed
        - in: query
          name: instanceName
          required: true
          schema:
            type: string
          description: Target instance tag/name
        - in: query
          name: clusterId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Route removal task dispatched
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          taskId:
                            type: string
                          message:
                            type: string
        '400':
          description: Missing instanceName or service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions (developer role required)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Node not connected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ─── Status (Wake-on-demand) ────────────────────────────────────────────────
...